proj
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/mipmap@5704 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CD3D9HLSLMaterialRenderer.cpp
@@ -91,7 +91,7 @@ bool CD3D9HLSLMaterialRenderer::createHLSLVertexShader(const char* vertexShaderP
 	// compile without debug info
 	HRESULT h = stubD3DXCompileShader(
 		vertexShaderProgram,
-		strlen(vertexShaderProgram),
+		(UINT)strlen(vertexShaderProgram),
 		0, // macros
 		0, // no includes
 		shaderEntryPointName,
